Make MAX_SIZE a uint64_t .
Summary: It's only ever used as a uint64_t anyway.

Test Plan:
  make check

Reviewers: freetrader, sickpig, awemany, kyuupichan, dagurval, #bitcoin_abc

Reviewed By: dagurval, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D338

+++ src/serialize.h
@@ -23,7 +23,7 @@
 
 #include "prevector.h"
 
-static const unsigned int MAX_SIZE = 0x02000000;
+static const uint64_t MAX_SIZE = 0x02000000;
 
 /**
  * Dummy data type to identify deserializing constructors.
@@ -291,7 +291,7 @@ template <typename Stream> uint64_t ReadCompactSize(Stream &is) {
         if (nSizeRet < 0x100000000ULL)
             throw std::ios_base::failure("non-canonical ReadCompactSize()");
     }
-    if (nSizeRet > (uint64_t)MAX_SIZE) {
+    if (nSizeRet > MAX_SIZE) {
         throw std::ios_base::failure("ReadCompactSize(): size too large");
     }
     return nSizeRet;
